Initial commit.
Right now we have a running VM on Amazon EC2 called "master"
and our next goal is to have it create another VM, named "slave".

The tool that is supposed to be the foundation for this kind
of remote management is Ansible, which is capable to control AWS
using python "boto" interface. So we have installed boto, boto3
and botocore on "master".

The slave_start.yml is composed of Ansbile AWS tutorial sources
and it has not been tested yet.

Look for further updates in the branch "development".