The first attempt to build a complete system.
worker.py is meant to be copied to the slave
and run. It will accept http requests from the
master the respond with a random quote, which
will eventually get sent to the user and displayed.
The quote list is build of a number of sayings
from savvy people.

slave_start.yml:
Now it commands the slave to do what's described
above.

server.py:
/get_quote routing added - when received, the server
sends an http request to the slave to get a quote
and send it back to user.

slave_terminate.yml:
A minor fix - an unnecessary single whitespace removed.